WIP: Removing or uncommenting commented-out code; removing integ tests that relied on data from file; removed tests that relied on method ordering; s/BigInteger/Long/g for all entity ids.

+++ src/main/java/org/sgs/stashbot/dao/AuthTimeDao.java
@@ -4,11 +4,9 @@ import org.sgs.stashbot.model.AuthPollingTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 
 @Repository
-public interface AuthTimeDao extends JpaRepository<AuthPollingTime, BigInteger> {
+public interface AuthTimeDao extends JpaRepository<AuthPollingTime, Long> {
     // Get most recent successfult auth time
     AuthPollingTime findFirstBySuccessIsTrueOrderByDateDesc();
 }
