+++ src/main/java/org/sgs/stashbot/dao/StashUrlDao.java
@@ -1,13 +1,10 @@
 package org.sgs.stashbot.dao;
 
-import java.math.BigInteger;
-import java.util.List;
-
 import org.sgs.stashbot.model.StashUrl;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface StashUrlDao extends JpaRepository<StashUrl, BigInteger> {
+public interface StashUrlDao extends JpaRepository<StashUrl, Long> {
 }
